Login and register refactor,extract from component,
