+++ src/app/auth/style.css
@@ -1,0 +1,3 @@
+.form_wrapper {
+  @apply flex justify-center p-10  w-screen h-screen max-w-[650px];
+}
