--- src/app/signin/style.css
@@ -1,3 +1,0 @@
-.form_wrapper {
-  @apply mx-auto flex justify-center p-6 w-full h-screen max-w-[650px];
-}
